requirements.txt
openai>=1.3.0
streamlit>=1.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
